Memory leak fix.
