+++ test/core/runtime/runtime_test_base.hpp
@@ -148,6 +148,7 @@ class RuntimeTestBase : public ::testing::Test {
             .value();
 
     auto module_repo = std::make_shared<runtime::ModuleRepositoryImpl>(
+        std::make_shared<runtime::RuntimeInstancesPool>(),
         upgrade_tracker,
         module_factory,
         std::make_shared<runtime::SingleModuleCache>());
